feat(security): Implementa autorização e API de Anotações

+++ Controllers/FavoriteGamesController.cs
@@ -1,8 +1,10 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using SuppGamesBack.Data;
 using SuppGamesBack.Models;
 using SuppGamesBack.Models.DTOs;
 using SuppGamesBack.Services;
+using System.Security;
 
 namespace SuppGamesBack.Controllers
 {
@@ -11,8 +13,18 @@ namespace SuppGamesBack.Controllers
     public class FavoriteGamesController : ControllerBase
     {
         private readonly IFavoriteGameRepository _favoriteGameRepository;
-        private readonly IUserRepository _userRepository;
         private readonly IRawgClient _rawgClient;
+        private int? GetCurrentUser()
+        {
+            var userIdString = User.FindFirst(System.Security.Claims.ClaimTypes.NameIdentifier)?.Value;
+
+            if (int.TryParse(userIdString, out var userId))
+            {
+                return userId;
+            }
+
+            return null;
+        }
 
         public FavoriteGamesController(
             IFavoriteGameRepository favoriteGameRepository,
@@ -20,20 +32,21 @@ namespace SuppGamesBack.Controllers
             IRawgClient rawgClient)
         {
             _favoriteGameRepository = favoriteGameRepository;
-            _userRepository = userRepository;
             _rawgClient = rawgClient;
         }
 
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> AddFavoriteGame([FromBody] CreateFavoriteGameDTO favoriteDto)
         {
-            var user = await _userRepository.GetByIdAsync(favoriteDto.UserId);
-            if (user == null || !user.IsAtivo)
+
+            var userIdFromToken = GetCurrentUser();
+            if (userIdFromToken == null)
             {
-                return NotFound("Usuário não encontrado ou inativo.");
+                return Unauthorized();
             }
 
-            if (await _favoriteGameRepository.ExistsAsync(favoriteDto.UserId, favoriteDto.Slug))
+            if (await _favoriteGameRepository.ExistsAsync(userIdFromToken.Value, favoriteDto.Slug))
             {
                 return Conflict("Este jogo já está na sua lista de favoritos."); 
             }
@@ -48,7 +61,7 @@ namespace SuppGamesBack.Controllers
             {
                 Name = rawgGame.name,
                 Slug = rawgGame.slug,
-                UserId = favoriteDto.UserId,
+                UserId = userIdFromToken.Value,
                 IsFavorite = true,
                 CreateDate = DateTime.UtcNow,
                 LastUpdate = DateTime.UtcNow,
@@ -79,21 +92,24 @@ namespace SuppGamesBack.Controllers
             return CreatedAtAction(nameof(GetFavoriteGameById), new { id = createdGame.Id }, responseDto);
         }
 
-        [HttpGet("user/{userId}")]
-        public async Task<IActionResult> GetFavoriteGamesByUserId(int userId)
+        [HttpGet]
+        [Authorize]
+
+        public async Task<IActionResult> GetMyFavoriteGames()
         {
-            var user = await _userRepository.GetByIdAsync(userId);
-            if (user == null)
+            var userIdFromToken = GetCurrentUser();
+
+            if(userIdFromToken == null)
             {
-                return NotFound("Usuário não encontrado.");
+                return Unauthorized();
             }
 
-            var games = await _favoriteGameRepository.GetByUserIdAsync(userId);
+            var games = await _favoriteGameRepository.GetByUserIdAsync(userIdFromToken.Value);
 
             var responseDtos = games.Select(game => new FavoriteGameResponseDTO
             {
                 Id = game.Id,
-                UserId = game.UserId,
+                UserId = userIdFromToken.Value,
                 Name = game.Name,
                 Slug = game.Slug,
                 ImageUrl = game.ImageUrl,
@@ -106,13 +122,29 @@ namespace SuppGamesBack.Controllers
         }
 
         [HttpGet("{id}")]
+        [Authorize]
+
         public async Task<IActionResult> GetFavoriteGameById(int id)
         {
+            var userIdFromToken = GetCurrentUser();
+
             var game = await _favoriteGameRepository.GetByIdAsync(id);
+
+            if(userIdFromToken == null)
+            {
+                return Unauthorized();
+            }
+
             if (game == null)
             {
                 return NotFound();
             }
+
+            if(game.UserId != userIdFromToken)
+            {
+                return Forbid();
+            }
+
 
             var responseDto = new FavoriteGameResponseDTO
             {
@@ -129,13 +161,29 @@ namespace SuppGamesBack.Controllers
         }
 
         [HttpDelete("{id}")]
+        [Authorize]
+
         public async Task<IActionResult> DeleteFavoriteGame(int id)
         {
-            var success = await _favoriteGameRepository.DeleteAsync(id);
-            if (!success)
+            var userIdFromToken = GetCurrentUser();
+            var gameToDelete = await _favoriteGameRepository.GetByIdAsync(id);
+
+            if(userIdFromToken == null)
             {
-                return NotFound("Registro de jogo favorito não encontrado.");
+                return Unauthorized();
             }
+
+            if(gameToDelete == null)
+            {
+                return NotFound("Registro de jogo não encontrado.");
+            }
+
+            if(gameToDelete.UserId != userIdFromToken)
+            {
+                return Forbid();
+            }
+
+            await _favoriteGameRepository.DeleteAsync(id);
 
             return NoContent();
         }
